Break filter code out into module.
