+++ public/js/sortfilter.js
@@ -1,5 +1,4 @@
 var price_buckets = [.25, .5, 1, 2, 3, 4, 5, 7, 10, 15, 20, 25, 30, 40, 50, 75, 100];
-var rarity_order = ['common', 'uncommon', 'rare', 'mythic'];
 
 function GetColorCategory(type, colors) {
   if (type.toLowerCase().includes('land')) {
@@ -29,287 +28,6 @@ function GetColorCategory(type, colors) {
         return 'Colorless';
         break;
     }
-  }
-}
-
-function filterCard(card, filters) {
-
-  if (filters.length == 1) {
-    if (filters[0].type == 'token') {
-      return filterApply(card, filters[0]);
-    } else {
-      return filterCard(card, filters[0]);
-    }
-  } else {
-    if (filters.type == 'or') {
-      return (filters[0].type == 'token' ? filterApply(card, filters[0]) : filterCard(card, filters[0])) || (filters[1].type == 'token' ? filterApply(card, filters[1]) : filterCard(card, filters[1]))
-    } else {
-      return (filters[0].type == 'token' ? filterApply(card, filters[0]) : filterCard(card, filters[0])) && (filters[1].type == 'token' ? filterApply(card, filters[1]) : filterCard(card, filters[1]))
-    }
-  }
-
-}
-
-function areArraysEqualSets(a1, a2) {
-  if (a1.length != a2.length) return false;
-  let superSet = {};
-  for (let i = 0; i < a1.length; i++) {
-    const e = a1[i] + typeof a1[i];
-    superSet[e] = 1;
-  }
-
-  for (let i = 0; i < a2.length; i++) {
-    const e = a2[i] + typeof a2[i];
-    if (!superSet[e]) {
-      return false;
-    }
-    superSet[e] = 2;
-  }
-
-  for (let e in superSet) {
-    if (superSet[e] === 1) {
-      return false;
-    }
-  }
-
-  return true;
-}
-
-function arrayContainsOtherArray(arr1, arr2) {
-  return arr2.every(v => arr1.includes(v));
-}
-
-
-function filterApply(card, filter) {
-  let res = null;
-  if (filter.category == 'name') {
-    res = card.details.name_lower.indexOf(filter.arg) > -1;
-  }
-  if (filter.category == 'oracle' && card.details.oracle_text) {
-    res = card.details.oracle_text.toLowerCase().indexOf(filter.arg) > -1;
-  }
-  if (filter.category == 'color' && card.details.colors) {
-    switch (filter.operand) {
-      case ':':
-      case '=':
-        if (filter.arg.length == 1 && filter.arg[0] == 'C') {
-          res = !card.details.colors.length;
-        } else {
-          res = areArraysEqualSets(card.details.colors, filter.arg);
-        }
-        break;
-      case '<':
-        res = arrayContainsOtherArray(filter.arg, card.details.colors) && card.details.colors.length < filter.arg.length;
-        break;
-      case '>':
-        res = arrayContainsOtherArray(card.details.colors, filter.arg) && card.details.colors.length > filter.arg.length;
-        break;
-      case '<=':
-        res = arrayContainsOtherArray(filter.arg, card.details.colors) && card.details.colors.length <= filter.arg.length;
-        break;
-      case '>=':
-        res = arrayContainsOtherArray(card.details.colors, filter.arg) && card.details.colors.length >= filter.arg.length;
-        break;
-    }
-  }
-  if (filter.category == 'identity' && card.colors) {
-    switch (filter.operand) {
-      case ':':
-      case '=':
-        if (filter.arg.length == 1 && filter.arg[0] == 'C') {
-          res = !card.colors.length;
-        } else {
-          res = areArraysEqualSets(card.colors, filter.arg);
-        }
-        break;
-      case '<':
-        res = arrayContainsOtherArray(filter.arg, card.colors) && card.details.color_identity.length < filter.arg.length;
-        break;
-      case '>':
-        res = arrayContainsOtherArray(card.colors, filter.arg) && card.details.color_identity.length > filter.arg.length;
-        break;
-      case '<=':
-        res = arrayContainsOtherArray(filter.arg, card.colors) && card.details.color_identity.length <= filter.arg.length;
-        break;
-      case '>=':
-        res = arrayContainsOtherArray(card.colors, filter.arg) && card.details.color_identity.length >= filter.arg.length;
-        break;
-    }
-  }
-  if (filter.category == 'mana' && card.details.parsed_cost) {
-    res = areArraysEqualSets(card.details.parsed_cost, filter.arg);
-  }
-  if (filter.category == 'cmc' && card.cmc) {
-    switch (filter.operand) {
-      case ':':
-      case '=':
-        res = filter.arg == card.cmc;
-        break;
-      case '<':
-        res = card.cmc < filter.arg;
-        break;
-      case '>':
-        res = card.cmc > filter.arg;
-        break;
-      case '<=':
-        res = card.cmc <= filter.arg;
-        break;
-      case '>=':
-        res = card.cmc >= filter.arg;
-        break;
-    }
-  }
-  if (filter.category == 'type' && card.details.type) {
-    if (card.details.type.toLowerCase().indexOf(filter.arg) > -1) {
-      res = true;
-    }
-  }
-  if (filter.category == 'power') {
-    if (card.details.power) {
-      switch (filter.operand) {
-        case ':':
-        case '=':
-          res = filter.arg == card.details.power;
-          break;
-        case '<':
-          res = card.details.power < filter.arg;
-          break;
-        case '>':
-          res = card.details.power > filter.arg;
-          break;
-        case '<=':
-          res = card.details.power <= filter.arg;
-          break;
-        case '>=':
-          res = card.details.power >= filter.arg;
-          break;
-      }
-    }
-  }
-  if (filter.category == 'toughness') {
-    if (card.details.toughness) {
-      switch (filter.operand) {
-        case ':':
-        case '=':
-          res = filter.arg == card.details.toughness;
-          break;
-        case '<':
-          res = card.details.toughness < filter.arg;
-          break;
-        case '>':
-          res = card.details.toughness > filter.arg;
-          break;
-        case '<=':
-          res = card.details.toughness <= filter.arg;
-          break;
-        case '>=':
-          res = card.details.toughness >= filter.arg;
-          break;
-      }
-    }
-  }
-  if (filter.category == 'tag') {
-    res = card.tags.some(element => {
-      return element.toLowerCase() == filter.arg;
-    });
-  }
-  if (filter.category == 'status') {
-    if (card.status.toLowerCase() == filter.arg) {
-      res = true;
-    }
-  }
-
-  if (filter.category == 'price') {
-    var price = null;
-    if (card.details.price) {
-      price = card.details.price;
-    } else if (card.details.price_foil) {
-      price = card.details.price_foil;
-    }
-    if (price) {
-      switch (filter.operand) {
-        case ':':
-        case '=':
-          res = filter.arg == price;
-          break;
-        case '<':
-          res = price < filter.arg;
-          break;
-        case '>':
-          res = price > filter.arg;
-          break;
-        case '<=':
-          res = price <= filter.arg;
-          break;
-        case '>=':
-          res = price >= filter.arg;
-          break;
-      }
-    }
-  }
-  if (filter.category == 'pricefoil') {
-    var price = card.details.price_foil || null
-    if (price) {
-      switch (filter.operand) {
-        case ':':
-        case '=':
-          res = filter.arg == price;
-          break;
-        case '<':
-          res = price < filter.arg;
-          break;
-        case '>':
-          res = price > filter.arg;
-          break;
-        case '<=':
-          res = price <= filter.arg;
-          break;
-        case '>=':
-          res = price >= filter.arg;
-          break;
-      }
-    }
-  }
-  if (filter.category == 'rarity') {
-    let rarity = card.details.rarity;
-    switch (filter.operand) {
-      case ':':
-      case '=':
-        res = rarity == filter.arg;
-        break;
-      case '<':
-        res = rarity_order.indexOf(rarity) < rarity_order.indexOf(filter.arg);
-        break;
-      case '>':
-        res = rarity_order.indexOf(rarity) > rarity_order.indexOf(filter.arg);
-        break;
-      case '<=':
-        res = rarity_order.indexOf(rarity) <= rarity_order.indexOf(filter.arg);
-        break;
-      case '>=':
-        res = rarity_order.indexOf(rarity) >= rarity_order.indexOf(filter.arg);
-        break;
-    }
-  }
-
-  if (filter.not) {
-    return !res;
-  } else {
-    return res;
-  }
-}
-
-
-//returns the price bucket label at the index designating the upper bound
-//at index == 0, returns < lowest
-//at index == length, returs >= highest
-function price_bucket_label(index) {
-  if (index == 0) {
-    return '< $' + price_buckets[0];
-  } else if (index == price_buckets.length) {
-    return '>= $' + price_buckets[price_buckets.length - 1];
-  } else {
-    return '$' + price_buckets[i - 1] + ' - $' + (price_buckets[i] - .01);
   }
 }
 
@@ -561,6 +279,19 @@ function cardIsLabel(card, label, sort) {
     } else {
       return label == "No Price Available";
     }
+  }
+}
+
+//returns the price bucket label at the index designating the upper bound
+//at index == 0, returns < lowest
+//at index == length, returs >= highest
+function price_bucket_label(index) {
+  if (index == 0) {
+    return '< $' + price_buckets[0];
+  } else if (index == price_buckets.length) {
+    return '>= $' + price_buckets[price_buckets.length - 1];
+  } else {
+    return '$' + price_buckets[i - 1] + ' - $' + (price_buckets[i] - .01);
   }
 }
